move number loading methods in init
